Split NativeMethods in several module dependant native classes
