+++ application/Modules/Mouse/Producers/MouseClickEventProducer.cs
@@ -1,9 +1,8 @@
-using System;
 using System.Windows;
 using System.Windows.Input;
 using MORR.Modules.Mouse.Events;
 using MORR.Shared.Events.Queue;
-using MORR.Shared.Utility;
+using MORR.Shared.Hook;
 
 namespace MORR.Modules.Mouse.Producers
 {
@@ -12,20 +11,20 @@ namespace MORR.Modules.Mouse.Producers
     /// </summary>
     public class MouseClickEventProducer : DefaultEventQueue<MouseClickEvent>
     {
-        private readonly NativeMethods.MessageType[] listenedMessagesTypes = 
+        private readonly GlobalHook.MessageType[] listenedMessagesTypes = 
         {
-            NativeMethods.MessageType.WM_RBUTTONDOWN,
-            NativeMethods.MessageType.WM_LBUTTONDOWN,
-            NativeMethods.MessageType.WM_MBUTTONDOWN,
-            NativeMethods.MessageType.WM_RBUTTONDBLCLK,
-            NativeMethods.MessageType.WM_LBUTTONDBLCLK,
-            NativeMethods.MessageType.WM_MBUTTONDBLCLK,
-            NativeMethods.MessageType.WM_NCRBUTTONDOWN,
-            NativeMethods.MessageType.WM_NCLBUTTONDOWN,
-            NativeMethods.MessageType.WM_NCMBUTTONDOWN,
-            NativeMethods.MessageType.WM_NCRBUTTONDBLCLK,
-            NativeMethods.MessageType.WM_NCLBUTTONDBLCLK,
-            NativeMethods.MessageType.WM_NCMBUTTONDBLCLK
+            GlobalHook.MessageType.WM_RBUTTONDOWN,
+            GlobalHook.MessageType.WM_LBUTTONDOWN,
+            GlobalHook.MessageType.WM_MBUTTONDOWN,
+            GlobalHook.MessageType.WM_RBUTTONDBLCLK,
+            GlobalHook.MessageType.WM_LBUTTONDBLCLK,
+            GlobalHook.MessageType.WM_MBUTTONDBLCLK,
+            GlobalHook.MessageType.WM_NCRBUTTONDOWN,
+            GlobalHook.MessageType.WM_NCLBUTTONDOWN,
+            GlobalHook.MessageType.WM_NCMBUTTONDOWN,
+            GlobalHook.MessageType.WM_NCRBUTTONDBLCLK,
+            GlobalHook.MessageType.WM_NCLBUTTONDBLCLK,
+            GlobalHook.MessageType.WM_NCMBUTTONDBLCLK
         };
 
         public void StartCapture()
@@ -42,7 +41,7 @@ namespace MORR.Modules.Mouse.Producers
 
         private void MouseHookCallback(GlobalHook.HookMessage hookMessage)
         {
-            var messageType = (NativeMethods.MessageType) hookMessage.Type;
+            var messageType = (GlobalHook.MessageType) hookMessage.Type;
             var mouseAction = GetMouseAction(messageType);
             if (mouseAction == MouseAction.None)
             {
@@ -56,22 +55,22 @@ namespace MORR.Modules.Mouse.Producers
         }
 
 
-        private MouseAction GetMouseAction(NativeMethods.MessageType messageType)
+        private MouseAction GetMouseAction(GlobalHook.MessageType messageType)
         {
             return messageType switch
             {
-                NativeMethods.MessageType.WM_RBUTTONDOWN => MouseAction.RightClick,
-                NativeMethods.MessageType.WM_NCRBUTTONDOWN => MouseAction.RightClick,
-                NativeMethods.MessageType.WM_LBUTTONDOWN => MouseAction.LeftClick,
-                NativeMethods.MessageType.WM_NCLBUTTONDOWN => MouseAction.LeftClick,
-                NativeMethods.MessageType.WM_MBUTTONDOWN => MouseAction.MiddleClick,
-                NativeMethods.MessageType.WM_NCMBUTTONDOWN => MouseAction.MiddleClick,
-                NativeMethods.MessageType.WM_RBUTTONDBLCLK => MouseAction.RightDoubleClick,
-                NativeMethods.MessageType.WM_NCRBUTTONDBLCLK => MouseAction.RightDoubleClick,
-                NativeMethods.MessageType.WM_LBUTTONDBLCLK => MouseAction.LeftDoubleClick,
-                NativeMethods.MessageType.WM_NCLBUTTONDBLCLK => MouseAction.LeftDoubleClick,
-                NativeMethods.MessageType.WM_MBUTTONDBLCLK => MouseAction.MiddleDoubleClick,
-                NativeMethods.MessageType.WM_NCMBUTTONDBLCLK => MouseAction.MiddleDoubleClick,
+                GlobalHook.MessageType.WM_RBUTTONDOWN => MouseAction.RightClick,
+                GlobalHook.MessageType.WM_NCRBUTTONDOWN => MouseAction.RightClick,
+                GlobalHook.MessageType.WM_LBUTTONDOWN => MouseAction.LeftClick,
+                GlobalHook.MessageType.WM_NCLBUTTONDOWN => MouseAction.LeftClick,
+                GlobalHook.MessageType.WM_MBUTTONDOWN => MouseAction.MiddleClick,
+                GlobalHook.MessageType.WM_NCMBUTTONDOWN => MouseAction.MiddleClick,
+                GlobalHook.MessageType.WM_RBUTTONDBLCLK => MouseAction.RightDoubleClick,
+                GlobalHook.MessageType.WM_NCRBUTTONDBLCLK => MouseAction.RightDoubleClick,
+                GlobalHook.MessageType.WM_LBUTTONDBLCLK => MouseAction.LeftDoubleClick,
+                GlobalHook.MessageType.WM_NCLBUTTONDBLCLK => MouseAction.LeftDoubleClick,
+                GlobalHook.MessageType.WM_MBUTTONDBLCLK => MouseAction.MiddleDoubleClick,
+                GlobalHook.MessageType.WM_NCMBUTTONDBLCLK => MouseAction.MiddleDoubleClick,
                 _ => MouseAction.None
             };
         }
