+++ application/Modules/WindowManagement/Producers/WindowMovementEventProducer.cs
@@ -1,7 +1,8 @@
 ﻿using System.Drawing;
 using MORR.Modules.WindowManagement.Events;
+using MORR.Modules.WindowManagement.Native;
 using MORR.Shared.Events.Queue;
-using MORR.Shared.Utility;
+using MORR.Shared.Hook;
 using Point = System.Windows.Point;
 
 namespace MORR.Modules.WindowManagement.Producers
@@ -29,8 +30,8 @@ namespace MORR.Modules.WindowManagement.Producers
         /// </summary>
         private int windowUnderChangeHwnd;
 
-        private readonly NativeMethods.MessageType[] listenedMessageTypes =
-            { NativeMethods.MessageType.WM_ENTERSIZEMOVE, NativeMethods.MessageType.WM_EXITSIZEMOVE };
+        private readonly GlobalHook.MessageType[] listenedMessageTypes =
+            { GlobalHook.MessageType.WM_ENTERSIZEMOVE, GlobalHook.MessageType.WM_EXITSIZEMOVE };
 
         public void StartCapture()
         {
@@ -57,18 +58,18 @@ namespace MORR.Modules.WindowManagement.Producers
         /// <param name="msg">the hook message</param>
         private void WindowHookCallback(GlobalHook.HookMessage msg)
         {
-            if (msg.Type == (uint) NativeMethods.MessageType.WM_ENTERSIZEMOVE)
+            if (msg.Type == (uint) GlobalHook.MessageType.WM_ENTERSIZEMOVE)
             {
                 windowUnderChangeHwnd = (int) msg.Hwnd;
                 windowRecBeforeChange = new Rectangle();
-                NativeMethods.GetWindowRect(windowUnderChangeHwnd, ref windowRecBeforeChange);
+                NativeWindowManagement.GetWindowRect(windowUnderChangeHwnd, ref windowRecBeforeChange);
             }
 
-            if (msg.Type == (uint) NativeMethods.MessageType.WM_EXITSIZEMOVE)
+            if (msg.Type == (uint) GlobalHook.MessageType.WM_EXITSIZEMOVE)
             {
                 windowRecAfterChange = new Rectangle();
-                NativeMethods.GetWindowRect(windowUnderChangeHwnd, ref windowRecAfterChange);
-                if (Utility.IsRectSizeEqual(windowRecBeforeChange, windowRecAfterChange))
+                NativeWindowManagement.GetWindowRect(windowUnderChangeHwnd, ref windowRecAfterChange);
+                if (NativeWindowManagement.IsRectSizeEqual(windowRecBeforeChange, windowRecAfterChange))
                 {
                     var oldLocation = new Point { X = windowRecBeforeChange.X, Y = windowRecBeforeChange.Y };
                     var newLocation = new Point { X = windowRecAfterChange.X, Y = windowRecAfterChange.Y };
@@ -77,8 +78,8 @@ namespace MORR.Modules.WindowManagement.Producers
                         IssuingModule = WindowManagementModule.Identifier,
                         OldLocation = oldLocation,
                         NewLocation = newLocation,
-                        Title = Utility.GetWindowTitleFromHwnd(msg.Hwnd),
-                        ProcessName = Utility.GetProcessNameFromHwnd(msg.Hwnd)
+                        Title = NativeWindowManagement.GetWindowTitleFromHwnd(msg.Hwnd),
+                        ProcessName = NativeWindowManagement.GetProcessNameFromHwnd(msg.Hwnd)
                     };
                     Enqueue(@event);
                 }
