+++ application/Modules/WindowManagement/Producers/WindowStateChangedEventProducer.cs
@@ -1,8 +1,9 @@
 ﻿using System.Drawing;
 using System.Windows;
 using MORR.Modules.WindowManagement.Events;
+using MORR.Modules.WindowManagement.Native;
 using MORR.Shared.Events.Queue;
-using MORR.Shared.Utility;
+using MORR.Shared.Hook;
 
 namespace MORR.Modules.WindowManagement.Producers
 {
@@ -33,11 +34,11 @@ namespace MORR.Modules.WindowManagement.Producers
         /// </summary>
         private int windowUnderChangeHwnd;
 
-        private readonly NativeMethods.MessageType[] listenedMessageTypes =
-            { 
-                NativeMethods.MessageType.WM_SIZE,
-                NativeMethods.MessageType.WM_ENTERSIZEMOVE,
-                NativeMethods.MessageType.WM_EXITSIZEMOVE
+        private readonly GlobalHook.MessageType[] listenedMessageTypes =
+            {
+                GlobalHook.MessageType.WM_SIZE,
+                GlobalHook.MessageType.WM_ENTERSIZEMOVE,
+                GlobalHook.MessageType.WM_EXITSIZEMOVE
             };
 
         public void StartCapture()
@@ -55,14 +56,14 @@ namespace MORR.Modules.WindowManagement.Producers
         private void WindowHookCallback(GlobalHook.HookMessage msg)
         {
             // for detection of WindowState.Maximized and WindowState.Minimized
-            if (msg.Type == (uint) NativeMethods.MessageType.WM_SIZE &&
+            if (msg.Type == (uint) GlobalHook.MessageType.WM_SIZE &&
                 ((uint) msg.wParam == SIZE_MINIMIZED || (uint) msg.wParam == SIZE_MAXIMIZED))
             {
                 var @event = new WindowStateChangedEvent
                 {
                     IssuingModule = WindowManagementModule.Identifier,
-                    ProcessName = Utility.GetProcessNameFromHwnd(msg.Hwnd),
-                    Title = Utility.GetProcessNameFromHwnd(msg.Hwnd),
+                    ProcessName = NativeWindowManagement.GetProcessNameFromHwnd(msg.Hwnd),
+                    Title = NativeWindowManagement.GetProcessNameFromHwnd(msg.Hwnd),
                     // SIZE_MINIMIZED matches to the WindowState.Minimized in number
                     // SIZE_MAXIMIZED matches to the WindowState.Maximized in number
                     WindowState = (WindowState) msg.wParam
@@ -71,24 +72,24 @@ namespace MORR.Modules.WindowManagement.Producers
             }
 
             // for detection of WindowState.Normal (The window is restored.)
-            if (msg.Type == (uint) NativeMethods.MessageType.WM_ENTERSIZEMOVE)
+            if (msg.Type == (uint) GlobalHook.MessageType.WM_ENTERSIZEMOVE)
             {
                 windowUnderChangeHwnd = (int) msg.Hwnd;
                 windowRecBeforeChange = new Rectangle();
-                NativeMethods.GetWindowRect(windowUnderChangeHwnd, ref windowRecBeforeChange);
+                NativeWindowManagement.GetWindowRect(windowUnderChangeHwnd, ref windowRecBeforeChange);
             }
 
-            if (msg.Type == (uint) NativeMethods.MessageType.WM_EXITSIZEMOVE)
+            if (msg.Type == (uint) GlobalHook.MessageType.WM_EXITSIZEMOVE)
             {
                 windowRecAfterChange = new Rectangle();
-                NativeMethods.GetWindowRect(windowUnderChangeHwnd, ref windowRecAfterChange);
-                if (!Utility.IsRectSizeEqual(windowRecBeforeChange, windowRecAfterChange))
+                NativeWindowManagement.GetWindowRect(windowUnderChangeHwnd, ref windowRecAfterChange);
+                if (!NativeWindowManagement.IsRectSizeEqual(windowRecBeforeChange, windowRecAfterChange))
                 {
                     var @event = new WindowStateChangedEvent
                     {
                         IssuingModule = WindowManagementModule.Identifier,
-                        Title = Utility.GetWindowTitleFromHwnd(msg.Hwnd),
-                        ProcessName = Utility.GetProcessNameFromHwnd(msg.Hwnd),
+                        Title = NativeWindowManagement.GetWindowTitleFromHwnd(msg.Hwnd),
+                        ProcessName = NativeWindowManagement.GetProcessNameFromHwnd(msg.Hwnd),
                         WindowState = SIZE_RESTORED
                     };
                     Enqueue(@event);
